Fix: Cast sum() result to int for strict types compatibility
MySQL's SUM() returns string, but getQuantityForPeriod() declares
int return type. With declare(strict_types=1), this causes TypeError.

+++ app/Services/ProductBadgeService.php
@@ -114,7 +114,7 @@ readonly class ProductBadgeService
 
     private function getQuantityForPeriod(string $sku, DateRange $dateRange): int
     {
-        return OrderItem::query()
+        return (int) OrderItem::query()
             ->where('sku', $sku)
             ->whereHas('order', fn (Builder $query) => $query->whereBetween('received_at', [$dateRange->from, $dateRange->to])
             )
